Add files via upload
This is a neural network that could be trained to recognize anything that the user wants. To train the network, the user has to upload an excel file with a large set of training data. Each row in the file should start of with the desired result, following by data points that correspond to the desired result. The software will do the rest.
This neural network utilizes the sigmoid function for neuron activation. The network also uses the stochastic mode for batch sizes, meaning that it updates the weights and biases for every sample.